refactor: вынести код кнопки ещё в хук
